Teach Prism to stop polluting the console with AJV unknown format warnings
In ajv-validator/ajv#466, the author of AJV makes it very clear that people should be aware of unknown formats.

However, according to the OpenAPI spec, format is is an open string-valued property, and can have any value.

Given some traffic, all those warnings tend to pollute the console output a lot and make it very hard to follow up with what's logged by Prism.

This change intercepts AJV calls to console.warn and when they relate to unknown formats, silence them.