Removes `content-encoding` response header since we decompressed response
When a client requests endpoint with "Accept-Encoding: gzip" we fetched compressed response but node-fetch decopress content automatically so we can validate it. We send back decopressed response but at the same time we send original headers that contain "Content-Encoding: gzip". So the client gets Content-Encoding header and tries to decompress the content and gets an error since the content is already decompressed.

Perhaps we could go even further and return a compressed response (as the original server does), but I don't think that is necessary, at least for the moment.